project/webapp/requirements.txt

azure-storage-blob
Flask 
asyncio
azure-eventhub